requirements.txt
Flask==1.1.2
redis==3.5.3
PyMySQL==1.0.2
Flask_SQLAlchemy==2.5.1
SQLAlchemy==1.4.22
Pillow==8.3.1
